keygroup now implements ILevelMonitor interface.
Previously the IDE error checking compiler would crash when
there was a syntax or code error. The error checking compiler was unable
to pinpoint where the syntax or other error was. I believe the bug
occured because some code was split between the LucidityBaseClasses
package. (I can't remember exactly what that code was now that I've
removed the LucidityBaseClasses package.)