plugin parameter values are now set/get to/from the plugin parameter
manager. This allows for plugin parameters to be 'different' the value
the audio engine sees. This is a requirement for smoothing parameter
changes. The audio engire requires parameters to smoothing change over
time. While the GUI and other "views" require the parameter to always
reflect the most up to date value.