HACKED out the the knob smoother functionality.
This version is hacky because I've compared the file with the last
pre-smoother knob version and crudely erased the relevant differences.
The problem is there has been a number of changes since adding the
smoother that need to remain. I will need to test the knob to confirm
everything works.