Big Change: Changed the way the LFO frequency is calculated.
The LFO frequency has been changed to be much more similar to Zebra.
The LFO frequency now has a Max LFO rate and the rate knob linearly
scales between 0hz and Max Freq. This applies to the sync'ed LFO rates
too. Previously the sync'd lfos would snap to multples of the
sync rate.