Changed the ZeroOjbect Register() and Deregister() functions to help...
avoid deadlocks. Previously there would be a deadlock if a message
caused an ZeroObject to be added or removed from the mothership. The
potential for deadlocks still exists but has been reduced.

There will still be a deadlock if a VCL, Main or AudioMessage causes
an object of the same type to be added or removed from the mother ship.